add stereo pnp solver maybe better

+++ include/type.h
@@ -4,7 +4,7 @@
 #include <Eigen/Dense>
 namespace VISG {
 	using KeyPoints = std::vector<cv::KeyPoint>;
-	using MatchPoints = std::vector<Eigen::Vector3f>;
+	using MatchPoints = std::vector<Eigen::Vector3d>;
 	using MyMatches = std::vector<std::pair<int, int>>;
 }
